refactor: extract edit mode keyboard handling to separate hook
Move keyboard event handling from useEditMode to a new useEditModeKeyboard
hook to avoid duplicate event listeners when the hook is used in multiple
components. The new hook should be called ONCE at the top level
(InterfaceManager).

Changes:
- Add useEditModeKeyboard hook with hold-to-toggle and escape handling
- Move isHolding and holdProgress state to editStore for global access
- Update useEditMode to read state from store instead of local state
- Call useEditModeKeyboard in DesktopInterfaceManager
